tinsert st update|st fix

+++ test/tilelang_st/npu/a5/src/st/testcase/tinsert/launch.cpp
@@ -15,7 +15,7 @@
 extern "C" __global__ AICORE void TINSERT_acc2mat_f16_16x16(
     __gm__ uint16_t *a, __gm__ uint16_t *b, __gm__ uint16_t *c);
 extern "C" __global__ AICORE void TINSERT_acc2mat_f16_32x32(
-    __gm__ uint16_t *a, __gm__ uint16_t *b, __gm__ uint16_t *c);
+    __gm__ uint16_t *a, __gm__ uint16_t *b, __gm__ uint32_t *c);
 extern "C" __global__ AICORE void TINSERT_acc2mat_bf16_16x16(
     __gm__ uint16_t *a, __gm__ uint16_t *b, __gm__ uint16_t *c);
 extern "C" __global__ AICORE void TINSERT_acc2mat_f32_16x16(
@@ -31,8 +31,8 @@ void LaunchAcc2Mat_f16_16x16(uint16_t *a, uint16_t *b, uint16_t *c, void *stream
     TINSERT_acc2mat_f16_16x16<<<1, nullptr, stream>>>((__gm__ uint16_t *)a, (__gm__ uint16_t *)b, (__gm__ uint16_t *)c);
 }
 
-void LaunchAcc2Mat_f16_32x32(uint16_t *a, uint16_t *b, uint16_t *c, void *stream) {
-    TINSERT_acc2mat_f16_32x32<<<1, nullptr, stream>>>((__gm__ uint16_t *)a, (__gm__ uint16_t *)b, (__gm__ uint16_t *)c);
+void LaunchAcc2Mat_f16_32x32(uint16_t *a, uint16_t *b, uint32_t *c, void *stream) {
+    TINSERT_acc2mat_f16_32x32<<<1, nullptr, stream>>>((__gm__ uint16_t *)a, (__gm__ uint16_t *)b, (__gm__ uint32_t *)c);
 }
 
 void LaunchAcc2Mat_bf16_16x16(uint16_t *a, uint16_t *b, uint16_t *c, void *stream) {
